+++ src/main/java/net/p3pp3rf1y/sophisticatedbackpacks/backpack/wrapper/BackpackInventoryHandler.java
@@ -12,12 +12,12 @@ import net.p3pp3rf1y.sophisticatedcore.inventory.InventoryHandler;
 
 public class BackpackInventoryHandler extends InventoryHandler {
 	public BackpackInventoryHandler(int numberOfInventorySlots, IStorageWrapper storageWrapper, CompoundTag contentsNbt, Runnable saveHandler, int slotLimit) {
-		super(numberOfInventorySlots, storageWrapper, contentsNbt, saveHandler, slotLimit, Config.COMMON.stackUpgrade);
+		super(numberOfInventorySlots, storageWrapper, contentsNbt, saveHandler, slotLimit, Config.SERVER.stackUpgrade);
 	}
 
 	@Override
 	protected boolean isAllowed(ItemVariant resource) {
-		return !Config.COMMON.disallowedItems.isItemDisallowed(resource.getItem())
+		return !Config.SERVER.disallowedItems.isItemDisallowed(resource.getItem())
 				&& (!(resource.getItem() instanceof BackpackItem) || (hasInceptionUpgrade() && isBackpackWithoutInceptionUpgrade(resource.toStack())));
 	}