+++ src/main/java/net/p3pp3rf1y/sophisticatedbackpacks/client/KeybindHandler.java
@@ -2,6 +2,7 @@ package net.p3pp3rf1y.sophisticatedbackpacks.client;
 
 import com.mojang.blaze3d.platform.InputConstants;
 
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenKeyboardEvents;
@@ -17,7 +18,6 @@ import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.Container;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.phys.BlockHitResult;
@@ -36,8 +36,6 @@ import net.p3pp3rf1y.sophisticatedbackpacks.network.InventoryInteractionMessage;
 import net.p3pp3rf1y.sophisticatedbackpacks.network.SBPPacketHandler;
 import net.p3pp3rf1y.sophisticatedbackpacks.network.UpgradeToggleMessage;
 import net.p3pp3rf1y.sophisticatedbackpacks.util.PlayerInventoryProvider;
-import net.p3pp3rf1y.sophisticatedcore.client.gui.utils.GuiHelper;
-import net.p3pp3rf1y.sophisticatedcore.event.client.ClientRawInputEvent;
 
 import java.util.Map;
 import java.util.Optional;
@@ -57,16 +55,12 @@ public class KeybindHandler {
 	private static final String KEYBIND_SOPHISTICATEDBACKPACKS_CATEGORY = "keybind.sophisticatedbackpacks.category";
 	public static final KeyMapping BACKPACK_TOGGLE_UPGRADE_5 = new KeyMapping(SBPTranslationHelper.INSTANCE.translKeybind("toggle_upgrade_5"),
 			InputConstants.Type.KEYSYM.getOrCreate(KEY_UNKNOWN).getValue(), KEYBIND_SOPHISTICATEDBACKPACKS_CATEGORY);
-
 	public static final KeyMapping BACKPACK_TOGGLE_UPGRADE_4 = new KeyMapping(SBPTranslationHelper.INSTANCE.translKeybind("toggle_upgrade_4"),
 		    InputConstants.Type.KEYSYM.getOrCreate(KEY_UNKNOWN).getValue(), KEYBIND_SOPHISTICATEDBACKPACKS_CATEGORY);
-
 	public static final KeyMapping BACKPACK_TOGGLE_UPGRADE_3 = new KeyMapping(SBPTranslationHelper.INSTANCE.translKeybind("toggle_upgrade_3"),
 		    InputConstants.Type.KEYSYM.getOrCreate(KEY_UNKNOWN).getValue(), KEYBIND_SOPHISTICATEDBACKPACKS_CATEGORY);
-
 	public static final KeyMapping BACKPACK_TOGGLE_UPGRADE_2 = new KeyMapping(SBPTranslationHelper.INSTANCE.translKeybind("toggle_upgrade_2"),
 		    InputConstants.Type.KEYSYM.getOrCreate(KEY_X).getValue(), KEYBIND_SOPHISTICATEDBACKPACKS_CATEGORY);
-
 	public static final KeyMapping BACKPACK_TOGGLE_UPGRADE_1 = new KeyMapping(SBPTranslationHelper.INSTANCE.translKeybind("toggle_upgrade_1"),
 		    InputConstants.Type.KEYSYM.getOrCreate(KEY_Z).getValue(), KEYBIND_SOPHISTICATEDBACKPACKS_CATEGORY);
 
@@ -88,12 +82,14 @@ public class KeybindHandler {
 			InputConstants.Type.KEYSYM.getOrCreate(KEY_B).getValue(), KEYBIND_SOPHISTICATEDBACKPACKS_CATEGORY); // BackpackKeyConflictContext.INSTANCE
 
 	public static void register() {
-		ScreenEvents.BEFORE_INIT.register((client, screen, scaledWidth, scaledHeight) -> {
-			ScreenKeyboardEvents.allowKeyPress(screen).register(KeybindHandler::handleGuiKeyPress);
-			ScreenMouseEvents.allowMouseClick(screen).register(KeybindHandler::handleGuiMouseKeyPress);
-		});
+		if (!FabricLoader.getInstance().isModLoaded(CompatModIds.MKB)) {
+			ScreenEvents.BEFORE_INIT.register((client, screen, scaledWidth, scaledHeight) -> {
+				ScreenKeyboardEvents.allowKeyPress(screen).register(KeybindHandler::handleGuiKeyPress);
+				ScreenMouseEvents.allowMouseClick(screen).register(KeybindHandler::handleGuiMouseKeyPress);
+			});
+		}
 
-		ClientRawInputEvent.KEY_PRESSED.register(KeybindHandler::handleKeyInputEvent);
+		ClientTickEvents.END_CLIENT_TICK.register(KeybindHandler::handleKeyInputEvent);
 	}
 
 	public static void registerKeyMappings() {
@@ -106,7 +102,7 @@ public class KeybindHandler {
 	}
 
 	public static boolean handleGuiKeyPress(Screen screen, int key, int scancode, int modifiers) {
-		if (SORT_KEYBIND.isDown() && tryCallSort(screen)) {
+		if (SORT_KEYBIND.matches(key, scancode) && tryCallSort(screen)) {
 			return false;
 		} else if (BACKPACK_OPEN_KEYBIND.matches(key, scancode) && sendBackpackOpenOrCloseMessage()) {
 			return false;
@@ -125,33 +121,23 @@ public class KeybindHandler {
         return true;
 	}
 
-	public static InteractionResult handleKeyInputEvent(Minecraft minecraft, int key, int scancode, int action, int mods) {
+	private static void handleKeyInputEvent(Minecraft minecraft) {
 		if (BACKPACK_OPEN_KEYBIND.consumeClick()) {
 			sendBackpackOpenOrCloseMessage();
-			return InteractionResult.SUCCESS;
 		} else if (INVENTORY_INTERACTION_KEYBIND.consumeClick()) {
 			sendInteractWithInventoryMessage();
-			return InteractionResult.SUCCESS;
 		} else if (TOOL_SWAP_KEYBIND.consumeClick()) {
 			sendToolSwapMessage();
-			return InteractionResult.SUCCESS;
 		} else {
-			boolean success = false;
 			for (Map.Entry<Integer, KeyMapping> slotKeybind : UPGRADE_SLOT_TOGGLE_KEYBINDS.entrySet()) {
 				if (slotKeybind.getValue().consumeClick()) {
 					SBPPacketHandler.sendToServer(new UpgradeToggleMessage(slotKeybind.getKey()));
-					success = true;
 				}
 			}
-			if (success) {
-				return InteractionResult.SUCCESS;
-			}
 		}
-
-		return InteractionResult.PASS;
 	}
 
-	private static boolean tryCallSort(Screen gui) {
+	public static boolean tryCallSort(Screen gui) {
 		Minecraft mc = Minecraft.getInstance();
 		if (mc.player != null && mc.player.containerMenu instanceof BackpackContainer container && gui instanceof BackpackScreen screen) {
 			MouseHandler mh = mc.mouseHandler;
@@ -203,7 +189,7 @@ public class KeybindHandler {
 		SBPPacketHandler.sendToServer(new InventoryInteractionMessage(pos, blockraytraceresult.getDirection()));
 	}
 
-	private static boolean sendBackpackOpenOrCloseMessage() {
+	public static boolean sendBackpackOpenOrCloseMessage() {
 		if (Minecraft.getInstance().screen == null) {
 			SBPPacketHandler.sendToServer(new BackpackOpenMessage());
 			return false;
@@ -211,20 +197,19 @@ public class KeybindHandler {
 
 		Screen screen = Minecraft.getInstance().screen;
 		if (screen instanceof AbstractContainerScreen<?> containerScreen) {
-			return GuiHelper.getSlotUnderMouse(containerScreen).map(slot -> {
-				if (slot.container instanceof Inventory || isTrinket(slot.container)) {
-					Optional<PlayerInventoryReturn> ret = getPlayerInventory(slot);
-					if (ret.isPresent() && slot.getItem().getItem() instanceof BackpackItem) {
-						SBPPacketHandler.sendToServer(new BackpackOpenMessage(slot.getContainerSlot(), ret.get().identifier(), ret.get().handlerName()));
-						return true;
-					}
-				}
-				if (screen instanceof BackpackScreen && slot.getItem().getItem() instanceof BackpackItem && slot.getItem().getCount() == 1) {
-					SBPPacketHandler.sendToServer(new BackpackOpenMessage(slot.getContainerSlot()));
+			Slot slot = containerScreen.hoveredSlot;
+			if (slot != null && (slot.container instanceof Inventory || isTrinket(slot.container))) {
+				Optional<PlayerInventoryReturn> ret = getPlayerInventory(slot);
+
+				if (ret.isPresent() && slot.getItem().getItem() instanceof BackpackItem) {
+					SBPPacketHandler.sendToServer(new BackpackOpenMessage(slot.getContainerSlot(), ret.get().identifier(), ret.get().handlerName()));
 					return true;
 				}
-				return false;
-			}).orElse(false);
+			}
+			if (screen instanceof BackpackScreen && slot != null && slot.getItem().getItem() instanceof BackpackItem && slot.getItem().getCount() == 1) {
+				SBPPacketHandler.sendToServer(new BackpackOpenMessage(slot.getContainerSlot()));
+				return true;
+			}
 		}
 		return false;
 	}