+++ src/main/java/net/p3pp3rf1y/sophisticatedbackpacks/data/SBPLootInjectProvider.java
@@ -25,14 +25,14 @@ import java.util.Set;
 
 public class SBPLootInjectProvider implements DataProvider {
 	private static final String INJECT_FOLDER = "inject/";
-	public static final ResourceLocation ABANDONED_MINESHAFT = new ResourceLocation(SophisticatedBackpacks.ID, INJECT_FOLDER + BuiltInLootTables.ABANDONED_MINESHAFT.getPath());
-	public static final ResourceLocation BASTION_TREASURE = new ResourceLocation(SophisticatedBackpacks.ID, INJECT_FOLDER + BuiltInLootTables.BASTION_TREASURE.getPath());
-	public static final ResourceLocation DESERT_PYRAMID = new ResourceLocation(SophisticatedBackpacks.ID, INJECT_FOLDER + BuiltInLootTables.DESERT_PYRAMID.getPath());
-	public static final ResourceLocation END_CITY_TREASURE = new ResourceLocation(SophisticatedBackpacks.ID, INJECT_FOLDER + BuiltInLootTables.END_CITY_TREASURE.getPath());
-	public static final ResourceLocation NETHER_BRIDGE = new ResourceLocation(SophisticatedBackpacks.ID, INJECT_FOLDER + BuiltInLootTables.NETHER_BRIDGE.getPath());
-	public static final ResourceLocation SHIPWRECK_TREASURE = new ResourceLocation(SophisticatedBackpacks.ID, INJECT_FOLDER + BuiltInLootTables.SHIPWRECK_TREASURE.getPath());
-	public static final ResourceLocation SIMPLE_DUNGEON = new ResourceLocation(SophisticatedBackpacks.ID, INJECT_FOLDER + BuiltInLootTables.SIMPLE_DUNGEON.getPath());
-	public static final ResourceLocation WOODLAND_MANSION = new ResourceLocation(SophisticatedBackpacks.ID, INJECT_FOLDER + BuiltInLootTables.WOODLAND_MANSION.getPath());
+	public static final ResourceLocation ABANDONED_MINESHAFT = new ResourceLocation(SophisticatedBackpacks.MOD_ID, INJECT_FOLDER + BuiltInLootTables.ABANDONED_MINESHAFT.getPath());
+	public static final ResourceLocation BASTION_TREASURE = new ResourceLocation(SophisticatedBackpacks.MOD_ID, INJECT_FOLDER + BuiltInLootTables.BASTION_TREASURE.getPath());
+	public static final ResourceLocation DESERT_PYRAMID = new ResourceLocation(SophisticatedBackpacks.MOD_ID, INJECT_FOLDER + BuiltInLootTables.DESERT_PYRAMID.getPath());
+	public static final ResourceLocation END_CITY_TREASURE = new ResourceLocation(SophisticatedBackpacks.MOD_ID, INJECT_FOLDER + BuiltInLootTables.END_CITY_TREASURE.getPath());
+	public static final ResourceLocation NETHER_BRIDGE = new ResourceLocation(SophisticatedBackpacks.MOD_ID, INJECT_FOLDER + BuiltInLootTables.NETHER_BRIDGE.getPath());
+	public static final ResourceLocation SHIPWRECK_TREASURE = new ResourceLocation(SophisticatedBackpacks.MOD_ID, INJECT_FOLDER + BuiltInLootTables.SHIPWRECK_TREASURE.getPath());
+	public static final ResourceLocation SIMPLE_DUNGEON = new ResourceLocation(SophisticatedBackpacks.MOD_ID, INJECT_FOLDER + BuiltInLootTables.SIMPLE_DUNGEON.getPath());
+	public static final ResourceLocation WOODLAND_MANSION = new ResourceLocation(SophisticatedBackpacks.MOD_ID, INJECT_FOLDER + BuiltInLootTables.WOODLAND_MANSION.getPath());
 	public static final Set<ResourceLocation> ALL_TABLES = Set.of(ABANDONED_MINESHAFT, BASTION_TREASURE, DESERT_PYRAMID, END_CITY_TREASURE, NETHER_BRIDGE, SHIPWRECK_TREASURE, SIMPLE_DUNGEON, WOODLAND_MANSION);
 
 	private final DataGenerator generator;
@@ -45,17 +45,18 @@ public class SBPLootInjectProvider implements DataProvider {
 	public void run(CachedOutput cache) throws IOException {
 		Map<ResourceLocation, LootTable.Builder> tables = new HashMap<>();
 
-		tables.put(SIMPLE_DUNGEON, getLootTable(92,
-				getItemLootEntry(ModItems.BACKPACK, 4),
-				getItemLootEntry(ModItems.IRON_BACKPACK, 2),
-				getItemLootEntry(ModItems.PICKUP_UPGRADE, 2)));
-		tables.put(ABANDONED_MINESHAFT, getLootTable(89,
+		tables.put(SIMPLE_DUNGEON, getLootTable(90,
 				getItemLootEntry(ModItems.BACKPACK, 5),
+				getItemLootEntry(ModItems.COPPER_BACKPACK, 3),
+				getItemLootEntry(ModItems.PICKUP_UPGRADE, 2)));
+		tables.put(ABANDONED_MINESHAFT, getLootTable(84,
+				getItemLootEntry(ModItems.BACKPACK, 7),
+				getItemLootEntry(ModItems.COPPER_BACKPACK, 5),
 				getItemLootEntry(ModItems.IRON_BACKPACK, 3),
 				getItemLootEntry(ModItems.GOLD_BACKPACK, 1),
 				getItemLootEntry(ModItems.MAGNET_UPGRADE, 2)));
 		tables.put(DESERT_PYRAMID, getLootTable(89,
-				getItemLootEntry(ModItems.BACKPACK, 5),
+				getItemLootEntry(ModItems.COPPER_BACKPACK, 5),
 				getItemLootEntry(ModItems.IRON_BACKPACK, 3),
 				getItemLootEntry(ModItems.GOLD_BACKPACK, 1),
 				getItemLootEntry(ModItems.MAGNET_UPGRADE, 2)));
@@ -92,7 +93,7 @@ public class SBPLootInjectProvider implements DataProvider {
 	}
 
 	private static Path getPath(Path root, ResourceLocation id) {
-		return root.resolve("data/" + SophisticatedBackpacks.ID + "/loot_tables/" + id.getPath() + ".json");
+		return root.resolve("data/" + SophisticatedBackpacks.MOD_ID + "/loot_tables/" + id.getPath() + ".json");
 	}
 
 	private LootPoolEntryContainer.Builder<?> getItemLootEntry(Item item, int weight) {