+++ changelog.md
@@ -1,4 +1,12 @@
 # Changelog
 
+### General
+- Update to version 3.20.2
+- Code cleanup
+
 ### Fixes
 - Fix a compatibility issue with repurposed structures
+- Sorting keybind was not working with keys only with mouse buttons
+- Trinket Tags should not be validated against tooltip predicates but validator predicates
+- Fix a bug where in certain situations the refill upgrade was able to delete items
+- Fix open backpack keybinding, it was not possible to properly bind it to a mouse key